Evaluating the Submitted Code Results and Comparing it with the Test Cases Output

+++ app/controllers/missions_controller.rb
@@ -123,37 +123,47 @@ end
   def compile_user_code
 
     submitted_code = params[:submitted_code]
-    test_case = TestCase.find_by_id(2)
+    puts params[:current_mission_id]
+    test_cases = TestCase.where(mission_id: params[:current_mission_id])
 
-    java_code = 'public class Code{
+
+    test_cases.each do |tc|
+
+      java_code = 'public class Code{
 
       ' + submitted_code + '
 
       public static void main(String args[]){
 
         Code c = new Code();
-        int x = c.' + test_case.input + ';
-        System.out.println(x);
+        int x = c.' + tc.input + ';
+
+        System.out.println(""+x);
 
       }
 
     }'
 
-    my_file = File.new("Code.java", "w+")
-    my_file.puts(java_code)
-    my_file.close
-    File.chmod(0777,"Code.java")
+      my_file = File.new("Code.java", "w+")
+      my_file.puts(java_code)
+      my_file.close
+      File.chmod(0777,"Code.java")
+      stdin, stdout, stderr = Open3.popen3('javac Code.java')
 
-    stdin, stdout, stderr = Open3.popen3('javac Code.java')
-    puts stderr.gets
+      puts stderr.gets
+      $result = `timeout 4s java Code`
 
-    result = `timeout 4s java Code`
+      if $result.chomp != tc.output
+        render :json =>  {'output':'Failure'}
+        return
+      end
 
+      File.delete("Code.java")
+      File.delete("Code.class")
 
-    File.delete("Code.java")
-    File.delete("Code.class")
+    end
 
-    render :json =>  {'output':result.chomp}
+    render :json =>  {'output':'Success'}
 
   end
